Move most Protobuf work to `Internal`
